ai-engine/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[project]
name = "aetherverse-ai-engine"
version = "0.1.0"
description = "AetherVerse AI 引擎 (Agent B)"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.110",
    "uvicorn[standard]>=0.27",
    "pydantic>=2.5",
    "pydantic-settings>=2.1",
    "httpx>=0.27",
    "openai>=1.12",
    "redis>=5.0",
    "aio-pika>=9.4",
    "pillow>=10.2",
    "structlog>=24.1",
    "jinja2>=3.1",
    "pyyaml>=6.0",
    "aetherverse-shared",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "ruff>=0.3",
]

[tool.ruff]
line-length = 120
target-version = "py311"

[tool.pytest.ini_options]
asyncio_mode = "auto"
